feat(ai api): add openai api file (WIP)
