+++ src/pages/index.tsx
@@ -2,6 +2,9 @@ import Head from "next/head";
 
 import IdeaInputForm from "@/components/IdeaInputForm";
 
+
+
+
 export default function Home() {
   return (
     <>
